# client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "AppController_getHealth",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["minFactory"]
      }
    },
    "/minrps/games/{id}": {
      "get": {
        "operationId": "getById",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MinRpsGameResponseDto" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        },
        "summary": "",
        "tags": ["minRPS Game"]
      },
      "delete": {
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID of ressource",
            "schema": { "example": "550e8400-e29b-41d4-a716-446655440000", "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        },
        "summary": "",
        "tags": ["minRPS Game"]
      }
    },
    "/minrps/games": {
      "get": {
        "operationId": "getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/MinRpsGameResponseDto" }
                }
              }
            }
          },
          "500": { "description": "Internal Server Error" }
        },
        "summary": "",
        "tags": ["minRPS Game"]
      },
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MinRPSGameRequestDto" }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MinRpsGameResponseDto" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        },
        "summary": "",
        "tags": ["minRPS Game"]
      }
    }
  },
  "info": {
    "title": "minFactory",
    "description": "minFactory OpenAPI specification. <br/> <a href=\"http://localhost:3000/openapi-json/\" target=\"_blank\">JSON specification</a>",
    "version": "v0.0.0",
    "contact": { "name": "minFactory", "url": "https://minFactory.com", "email": "" }
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "MinRpsGameResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Game name",
            "example": "Jon Doe Game",
            "minLength": 2,
            "maxLength": 32
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the game was created (UTC)",
            "example": "2025-10-21T18:45:30.000Z"
          },
          "id": {
            "type": "string",
            "description": "Unique game identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "format": "uuid"
          }
        },
        "required": ["name", "createdAt", "id"]
      },
      "MinRPSGameRequestDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "player name", "example": "Jon Doe" }
        },
        "required": ["name"]
      }
    }
  }
}